Back button stores snapshots between steps
Previously, going back from the force step to the drawing step would not remove added forces. Now, a snapshot is stored whenever the "next" button is pressed, and one is restored whenever the "back" button is pressed. This will make "undoing" steps like the force drawing simpler in the future.