Adding density-setting as a configurable parameter
Density-setting is now hidden unless the URL parameter 'densityStepEnabled' is passed. This means conditionally hiding the step in the bottom bar, and skipping it after step 3. I held off on introducing an elegant way to hide additional steps, but if we added a second configurable step it's probably worth looking into.