Updating label calculations
The average of a plate's field coordinates might not be on the plate itself. Now, the field on a plate nearest the center is labeled, so the label always falls on the plate. A test case is added to show why this is necessary.

Also, stabilization code is added to keep labels from jumping around.